requirements.txt
aiohttp
requests
psycopg2-binary>=2.9.3
requests==2.31.0
python-dotenv==1.0.1
tqdm>=4.65.0